requirements.txt
asgiref==3.2.8
Django==3.0.7
django-filter==2.3.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
Pillow==7.1.2
psycopg2-binary==2.8.5
python-dotenv==0.13.0
pytz==2020.1
sqlparse==0.3.1
